feat(summary): Refactor runner to split summary and metrics into model and logic
