+++ benchmark-tests/benches/test_lib_bench/entry_point/test_lib_bench_entry_point.rs
@@ -3,9 +3,9 @@ use std::hint::black_box;
 use benchmark_tests::assert::Assert;
 use gungraun::prelude::*;
 use gungraun::{Callgrind, EntryPoint, EventKind, FlamegraphConfig, ValgrindTool};
+use gungraun_runner::metrics::model::Metric;
 use gungraun_runner::runner::callgrind::hashmap_parser::SourcePath;
-use gungraun_runner::runner::metrics::Metric;
-use gungraun_runner::runner::summary::{BenchmarkSummary, ToolMetricSummary};
+use gungraun_runner::summary::model::{BenchmarkSummary, ToolMetricSummary};
 
 #[inline(never)]
 fn nested() -> u64 {
