+++ benchmark-tests/tests/test_tool/test_error_metrics_parser.rs
@@ -1,11 +1,11 @@
 use std::path::PathBuf;
 
 use gungraun_runner::api::{ErrorMetric, ValgrindTool};
-use gungraun_runner::runner::metrics::Metrics;
-use gungraun_runner::runner::summary::ToolMetrics;
+use gungraun_runner::metrics::model::Metrics;
 use gungraun_runner::runner::tool::error_metric_parser::ErrorMetricLogfileParser;
 use gungraun_runner::runner::tool::parser::Parser;
 use gungraun_runner::runner::tool::path::ToolOutputPathKind;
+use gungraun_runner::summary::model::ToolMetrics;
 use pretty_assertions::assert_eq;
 use rstest::rstest;
