+++ benchmark-tests/tests/util/common.rs
@@ -8,9 +8,9 @@ use std::sync::LazyLock;
 
 use anyhow::{Context, Result};
 use gungraun::ValgrindTool;
-use gungraun_runner::runner::summary::BaselineKind;
 use gungraun_runner::runner::tasks::ProcessHandler;
 use gungraun_runner::runner::tool::path::{ToolOutputPath, ToolOutputPathKind};
+use gungraun_runner::summary::model::BaselineKind;
 use pretty_assertions::assert_eq;
 use serde::{Deserialize, Serialize};
