+++ gungraun-runner/Cargo.toml
@@ -21,7 +21,7 @@ rust-version.workspace = true
 version.workspace = true
 
 [features]
-api = ["dep:serde"]
+api = ["dep:serde", "either-or-both/serde"]
 default = ["runner"]
 runner = [
   "api",
@@ -33,9 +33,8 @@ runner = [
   "dep:crossbeam",
   "dep:derive_more",
   "dep:env_logger",
-  "dep:either-or-both",
   "dep:glob",
-  "dep:indexmap",
+  "indexmap/serde",
   "dep:inferno",
   "dep:itertools",
   "dep:log",
@@ -54,8 +53,15 @@ runner = [
   "dep:tempfile",
   "dep:version-compare",
   "dep:which",
+  "either-or-both/indexmap",
 ]
-schema = ["default", "dep:schemars", "either-or-both?/schema"]
+summary = [
+  "api",
+  "dep:derive_more",
+  "dep:schemars",
+  "either-or-both/schema",
+  "indexmap/serde",
+]
 
 [package.metadata.docs.rs]
 all-features = true
@@ -73,14 +79,10 @@ clap = { workspace = true, optional = true, features = [
 colored = { workspace = true, optional = true }
 crossbeam = { workspace = true, optional = true }
 derive_more = { workspace = true, optional = true, features = ["as_ref"] }
-either-or-both = { workspace = true, optional = true, features = [
-  "indexmap",
-  "serde",
-  "schema",
-] }
+either-or-both = { workspace = true, optional = true }
 env_logger = { workspace = true, optional = true }
 glob = { workspace = true, optional = true }
-indexmap = { workspace = true, optional = true, features = ["serde"] }
+indexmap = { workspace = true, optional = true }
 inferno = { workspace = true, optional = true, default-features = false, features = [
   'multithreaded',
 ] }
