+++ gungraun-runner/src/lib.rs
@@ -4,13 +4,17 @@
 #![doc(test(attr(warn(unused))))]
 #![doc(test(attr(allow(unused_extern_crates))))]
 
-#[cfg(feature = "api")]
+#[cfg(any(feature = "api", feature = "summary"))]
 pub mod api;
 #[cfg(feature = "runner")]
 pub mod error;
+#[cfg(any(feature = "runner", feature = "summary"))]
+pub mod metrics;
 #[cfg(feature = "runner")]
 pub mod runner;
-#[cfg(feature = "runner")]
+#[cfg(any(feature = "api", feature = "summary"))]
 pub mod serde;
+#[cfg(any(feature = "runner", feature = "summary"))]
+pub mod summary;
 #[cfg(feature = "runner")]
 pub mod util;
