+++ gungraun-runner/src/runner/callgrind/flamegraph.rs
@@ -14,8 +14,8 @@ use tempfile::TempDir;
 use super::flamegraph_parser::{FlamegraphMap, FlamegraphParser};
 use super::parser::{CallgrindParser, CallgrindProperties, Sentinel};
 use crate::api::{self, EventKind, FlamegraphKind};
-use crate::runner::summary::{BaselineKind, BaselineName, FlamegraphSummaries, FlamegraphSummary};
 use crate::runner::tool::path::{ToolOutputPath, ToolOutputPathKind};
+use crate::summary::model::{BaselineKind, BaselineName, FlamegraphSummaries, FlamegraphSummary};
 
 type ParserOutput = Vec<(PathBuf, CallgrindProperties, FlamegraphMap)>;
