+++ gungraun-runner/src/runner/lib_bench.rs
@@ -14,7 +14,6 @@ use anyhow::Result;
 
 use super::format::{LibraryBenchmarkHeader, ListFormat, OutputFormat};
 use super::meta::Metadata;
-use super::summary::{BaselineKind, BaselineName, BenchmarkKind, BenchmarkSummary, SummaryOutput};
 use super::tool::config::ToolConfigs;
 use super::tool::path::{ToolOutputPath, ToolOutputPathKind};
 use super::tool::run::RunOptions;
@@ -27,6 +26,9 @@ use crate::runner::common::{
     BaselineAndSaveDataProcessor, BaselineDataProcessor, Baselines, BenchmarkDataProcessor,
     BenchmarkSummaries, CapturedOutput, Config, Groups, LoadBaselineDataProcessor, ModulePath,
     Runner, SaveBaselineDataProcessor,
+};
+use crate::summary::model::{
+    BaselineKind, BaselineName, BenchmarkKind, BenchmarkSummary, SummaryOutput,
 };
 
 /// Implements [`Benchmark`] to compare a [`LibBench`] against one baseline and save the new run as
