+++ schema-gen/Cargo.toml
@@ -15,7 +15,7 @@ version.workspace = true
 
 [dependencies]
 gungraun-runner = { path = "../gungraun-runner", version = "=0.19.2", default-features = false, features = [
-  "schema",
+  "summary",
 ] }
 schemars = { workspace = true, features = ["indexmap2"] }
 serde_json = { workspace = true }
